NaniReturnService: Loading into Naninovel from an SGB map (#20)
This PR introduces a new Naninovel service that listens to Unity's onSceneLoaded event, and uses that to re-enter a specific Naniscript.

The intent is when you want to go from SGB to Naninovel, you have SGB load into a specifically named map. Then, the NaniReturnService can listen for this map name, and when that map scene is loaded, return to a specified naniscript.