SGB Saves in Naninovel State (#29)
This PR implements a bridging service that automatically hooks into IMod to make SGB store & load it's save data in the current Naninovel state.

This means that SGB save files can be written and read from naninovel save files.

At the moment, it only supports storing one save file, but this is a pretty artificial implementation. Down the line we can support multiple, but for the purpose of Dome-King's usecase, we really only need one.