NaniReturnService uses Pre-Load Operations (#36)
This PR revises the NaniReturnService to use SGB Map Pre-Load operations rather than relying on the Unity Scene Management callbacks to take action.

This is good because it makes the hand-off from SGB to Nani a little more stable, and a little quicker, as scenes that the return service is listening for do not need to be fully loaded first.