SGB Project-specific Fonts (#47)
This PR revises the custom font implementation to function using a targeting system. When an SGB project is loaded, the IMod Naninovel config is referenced to see what font it should use for THAT specific project.

This has been implemented for Dome-King, because it needs a different kind of font for the EN version and the JP version.